Fix for duplicate capstone utils symbols
