+++ shared/utils/capstone-utils.hpp
@@ -129,13 +129,7 @@ auto findNth(const uint32_t* addr) {
     SAFE_ABORT_MSG("Could not find: %u call at: %p within: %i rets, within size: %zu!", nToRetOn, addr, retCount, szBytes);
 }
 
-std::optional<uint32_t*> blConv(cs_insn* insn) {
-    if (insn->id == ARM64_INS_BL) {
-        // BL is pc + (imm << 2)
-        return reinterpret_cast<uint32_t*>((insn->detail->arm64.operands[0].imm << 2) + insn->address);
-    }
-    return std::nullopt;
-}
+std::optional<uint32_t*> blConv(cs_insn* insn);
 
 template<uint32_t nToRetOn, bool includeR = false, int retCount = -1, size_t szBytes = 4096>
 requires ((nToRetOn >= 1 && (szBytes % 4) == 0))
@@ -147,13 +141,7 @@ auto findNthBl(const uint32_t* addr) {
     }
 }
 
-std::optional<uint32_t*> bConv(cs_insn* insn) {
-    if (insn->id == ARM64_INS_B) {
-        // B is pc + (imm << 2)
-        return reinterpret_cast<uint32_t*>((insn->detail->arm64.operands[0].imm << 2) + insn->address);
-    }
-    return std::nullopt;
-}
+std::optional<uint32_t*> bConv(cs_insn* insn);
 
 template<uint32_t nToRetOn, bool includeR = false, int retCount = -1, size_t szBytes = 4096>
 requires ((nToRetOn >= 1 && (szBytes % 4) == 0))
@@ -165,19 +153,7 @@ auto findNthB(const uint32_t* addr) {
     }
 }
 
-std::optional<std::tuple<uint32_t*, arm64_reg, uint32_t*>> pcRelConv(cs_insn* insn) {
-    using tup = std::tuple<uint32_t*, arm64_reg, uint32_t*>;
-    switch (insn->id) {
-        case ARM64_INS_ADR:
-        // ADR is just pc + imm
-        return tup{reinterpret_cast<uint32_t*>(insn->address), insn->detail->arm64.operands[0].reg, reinterpret_cast<uint32_t*>(insn->detail->arm64.operands[1].imm + insn->address)};
-        case ARM64_INS_ADRP:
-        // ADRP is (pc & 1:12(0)) + (imm << 12)
-        return tup{reinterpret_cast<uint32_t*>(insn->address), insn->detail->arm64.operands[0].reg, reinterpret_cast<uint32_t*>(((insn->address >> 12) << 12) + (insn->detail->arm64.operands[1].imm << 12))};
-        default:
-        return std::nullopt;
-    }
-}
+std::optional<std::tuple<uint32_t*, arm64_reg, uint32_t*>> pcRelConv(cs_insn* insn);
 
 template<uint32_t nToRetOn, int retCount = -1, size_t szBytes = 4096>
 requires ((nToRetOn >= 1 && (szBytes % 4) == 0))
@@ -185,20 +161,7 @@ auto findNthPcRel(const uint32_t* addr) {
     return findNth<nToRetOn, &pcRelConv, &insnMatch<>, retCount, szBytes>(addr).value();
 }
 
-std::optional<std::tuple<uint32_t*, arm64_reg, int64_t>> regMatchConv(cs_insn* match, arm64_reg toMatch) {
-    // We need 1 to 2 operands, match 1 to 2 to register, determine dst reg from incoming instruction
-    // For now, it's pretty common for add immediates, which have dst as first op, src 2nd, imm third
-    auto& arm = match->detail->arm64;
-    using tup = std::tuple<uint32_t*, arm64_reg, int64_t>;
-    switch (match->id) {
-        case ARM64_INS_ADD:
-            if (arm.operands[1].reg != toMatch) return std::nullopt;
-            return tup{reinterpret_cast<uint32_t*>(match->address), arm.operands[0].reg, arm.operands[2].imm};
-        // TODO: Add more conversions for instructions!
-        default:
-        return std::nullopt;
-    }
-}
+std::optional<std::tuple<uint32_t*, arm64_reg, int64_t>> regMatchConv(cs_insn* match, arm64_reg toMatch);
 
 template<uint32_t nToRetOn, int retCount = -1, size_t szBytes = 4096>
 requires ((nToRetOn >= 1 && (szBytes % 4) == 0))
